move Fov into math module
